renombrar archivo donde se aloja la tabla
